+++ ftw/linkchecker/tests/test_report_upload.py
@@ -1,10 +1,10 @@
 from ftw.builder import Builder
 from ftw.builder import create
-from ftw.linkchecker.command import checking_links
-from ftw.linkchecker.command.checking_links import get_file_name, extract_config
+from ftw.linkchecker.accumulator import Accumulator
+from ftw.linkchecker.plone_site import PloneSite
+from ftw.linkchecker.report_generating import ReportHandler
 from ftw.linkchecker.tests import MultiPageTestCase
-from ftw.linkchecker.tests.exemplar_data.exemplar_config import config_file
-from ftw.linkchecker.tests.helper_generating_excel import generate_test_data_excel_workbook
+from ftw.linkchecker.tests.helpers import ConfigurationMock
 from ftw.testbrowser import browsing
 from zope.component.hooks import setSite
 import transaction
@@ -13,17 +13,23 @@ import transaction
 class TestReportUploading(MultiPageTestCase):
 
     @browsing
-    def test_report_generated_is_uploaded_to_file_listing_block_if_config_given(self, browser):
+    def test_report_generated_is_uploaded_to_file_listing_block_if_config(
+            self, browser):
         # plone site 0 has an upload config
         setSite(self.plone_site_objs[0])
         page = create(Builder('sl content page').titled(u'Content Page'))
         file_listing_block = create(Builder('sl listingblock')
                                     .titled('File Listing Block')
                                     .within(page))
-        xlsx_file = generate_test_data_excel_workbook()
-        file_name = get_file_name()
-        _, _, _, upload_location = extract_config(config_file)
-        checking_links.upload_report_to_filelistingblock(upload_location, xlsx_file, file_name)
+
+        # Run linchecker features needed for getting a report
+        configuration = ConfigurationMock(
+                'In real I would be *args comming from argparser.')
+        site = PloneSite(self.plone_site_objs[0], configuration)
+        accumulator = Accumulator(site)
+        accumulator.discover_broken_links()
+        report_handler = ReportHandler(site, accumulator)
+        report_handler.upload_report()
         transaction.commit()
 
         browser.visit(file_listing_block)
@@ -31,7 +37,9 @@ class TestReportUploading(MultiPageTestCase):
         report_donwload_url = node.attrib.get('href', '')
 
         # Since we assert a link this is always lower case.
-        expected_report_download_url_part = 'plone/content-page/file-listing-block/{}/download'.format(file_name).lower()
+        expected_report_download_url_part = (
+            'plone/content-page/file-listing-block/{}/download'.format(
+                report_handler._file_name).lower())
 
         self.assertIn(expected_report_download_url_part,
                       report_donwload_url,
